- Added gsettings to prepare MATE desktop - Removed remote-hosts from repo - addedd gsettings.sh to exclude.lst

+++ deploy.sh
@@ -10,19 +10,24 @@ then
 fi
 /usr/bin/zip -r $DOT_FILE . -x@exclude.lst
 
-echo [+] Iterate over destinations
-
-declare -a arr=(`cat remote-hosts`)
-
-for i in "${arr[@]}"
-do
-   echo [+] Deploy on $i
-   home_dir=`ssh $i pwd`
-   scp $DOT_FILE $i:$home_dir/$DOT_FILE
-   echo [+] dot-files transfered
-   ssh $i $EXTRACT_CMD
-   echo [+] data extractat at $i:$home_dir
-done
+if [ -f remote-hosts ]
+then
+    echo [+] Iterate over destinations
+    
+    declare -a arr=(`cat remote-hosts`)
+    
+    for i in "${arr[@]}"
+    do
+       echo [+] Deploy on $i
+       home_dir=`ssh $i pwd`
+       scp $DOT_FILE $i:$home_dir/$DOT_FILE
+       echo [+] dot-files transfered
+       ssh $i $EXTRACT_CMD
+       echo [+] data extractat at $i:$home_dir
+    done
+else
+    echo [!] No 'remote-hosts' found, Skip!
+fi
 
 echo [+] Deploy local
 prev_dir=`pwd`
